Support for sets.
