+++ setup.py
@@ -127,6 +127,10 @@ ccassandra_ext = Extension('ccassandra',
                            sources=[
                                'ccassandra/ccassandra.cpp',
                                'ccassandra/cql_types.cpp',
+                               'ccassandra/cql_type_utils.cpp',
+                               'ccassandra/cql_type_list.cpp',
+                               'ccassandra/cql_type_tuple.cpp',
+                               'ccassandra/cql_type_set.cpp',
                                'ccassandra/cql_type_factory.cpp',
                            ])
