+++ tests/unit/ccassandra/loading.py
@@ -19,6 +19,8 @@ if os.path.exists(build_path):
     sys.path.append(build_path)
 
 
+import ccassandra
+
 try:
     import ccassandra
 except ImportError:
